refactor: extract server-dm.js from server.js
Move DM message handlers into dedicated server-dm module:
dm_list, dm_open, dm_typing, dm_send, dm_add_favorite,
dm_remove_favorite. Includes mate DM open and mate DM send.

server.js delegates DM messages via dmHandler.handleMessage().

server.js: 1752 -> 1580 lines. server-dm.js: 208 lines.
