+++ lib/server.js
@@ -11,6 +11,7 @@ var mates = require("./mates");
 var sessionSearch = require("./session-search");
 var serverAuth = require("./server-auth");
 var serverSkills = require("./server-skills");
+var serverDm = require("./server-dm");
 var serverAdmin = require("./server-admin");
 var serverSettings = require("./server-settings");
 
@@ -1077,195 +1078,22 @@ function createServer(opts) {
     return true;
   }
 
-  // --- DM message handler (server-level, cross-project) ---
+  // --- DM message handler (delegated to server-dm + server-mates inline) ---
+  var dmHandler = serverDm.attachDm({
+    users: users,
+    dm: dm,
+    mates: mates,
+    projects: projects,
+    pushModule: pushModule,
+    addProject: addProject,
+  });
+
   function handleDmMessage(ws, msg) {
+    if (dmHandler.handleMessage(ws, msg)) return;
+
+    // --- Mate handlers (remaining in server.js, extracted next) ---
     if (!users.isMultiUser() || !ws._clayUser) return;
     var userId = ws._clayUser.id;
-
-    if (msg.type === "dm_list") {
-      var dmList = dm.getDmList(userId);
-      // Enrich with user info
-      for (var i = 0; i < dmList.length; i++) {
-        var otherUser = users.findUserById(dmList[i].otherUserId);
-        if (otherUser) {
-          var p = otherUser.profile || {};
-          dmList[i].otherUser = {
-            id: otherUser.id,
-            displayName: p.name || otherUser.displayName || otherUser.username,
-            username: otherUser.username,
-            avatarStyle: p.avatarStyle || "thumbs",
-            avatarSeed: p.avatarSeed || otherUser.username,
-            avatarColor: p.avatarColor || "#7c3aed",
-            avatarCustom: p.avatarCustom || "",
-          };
-        }
-      }
-      // Include mates in the list
-      var mateCtx = mates.buildMateCtx(userId);
-      var mateList = mates.getAllMates(mateCtx);
-      ws.send(JSON.stringify({ type: "dm_list", dms: dmList, mates: mateList }));
-      return;
-    }
-
-    if (msg.type === "dm_open") {
-      if (!msg.targetUserId) return;
-
-      // Check if target is a mate
-      var mateCtx2 = mates.buildMateCtx(userId);
-      if (mates.isMate(mateCtx2, msg.targetUserId)) {
-        var mate = mates.getMate(mateCtx2, msg.targetUserId);
-        if (!mate) return;
-        // Ensure mate project is registered (survives server restarts)
-        var mateSlug2 = "mate-" + mate.id;
-        if (!projects.has(mateSlug2)) {
-          var mateDir2 = mates.getMateDir(mateCtx2, mate.id);
-          fs.mkdirSync(mateDir2, { recursive: true });
-          var mateName2 = (mate.profile && mate.profile.displayName) || mate.name || "New Mate";
-          addProject(mateDir2, mateSlug2, mateName2, null, mate.createdBy || userId, null, { isMate: true, mateDisplayName: mateName2 });
-        }
-        var mp = mate.profile || {};
-        ws.send(JSON.stringify({
-          type: "dm_history",
-          dmKey: "mate:" + mate.id,
-          messages: dm.loadHistory("mate:" + mate.id),
-          isMate: true,
-          projectSlug: mateSlug2,
-          targetUser: {
-            id: mate.id,
-            displayName: mp.displayName || mate.name || "New Mate",
-            username: mate.id,
-            avatarStyle: mp.avatarStyle || "bottts",
-            avatarSeed: mp.avatarSeed || mate.id,
-            avatarColor: mp.avatarColor || "#6c5ce7",
-            avatarCustom: mp.avatarCustom || "",
-            isMate: true,
-            primary: !!mate.primary,
-            mateStatus: mate.status,
-            seedData: mate.seedData || {},
-          },
-        }));
-        return;
-      }
-
-      var result = dm.openDm(userId, msg.targetUserId);
-      var targetUser = users.findUserById(msg.targetUserId);
-      var tp = targetUser ? (targetUser.profile || {}) : {};
-      ws.send(JSON.stringify({
-        type: "dm_history",
-        dmKey: result.dmKey,
-        messages: result.messages,
-        targetUser: targetUser ? {
-          id: targetUser.id,
-          displayName: tp.name || targetUser.displayName || targetUser.username,
-          username: targetUser.username,
-          avatarStyle: tp.avatarStyle || "thumbs",
-          avatarSeed: tp.avatarSeed || targetUser.username,
-          avatarColor: tp.avatarColor || "#7c3aed",
-          avatarCustom: tp.avatarCustom || "",
-        } : null,
-      }));
-      return;
-    }
-
-    if (msg.type === "dm_typing") {
-      // Relay typing indicator to DM partner
-      var dmKey = msg.dmKey;
-      if (!dmKey) return;
-      var parts = dmKey.split(":");
-      if (parts.indexOf(userId) === -1) return;
-      var targetId = parts[0] === userId ? parts[1] : parts[0];
-      projects.forEach(function (ctx) {
-        ctx.forEachClient(function (otherWs) {
-          if (otherWs === ws) return;
-          if (!otherWs._clayUser || otherWs._clayUser.id !== targetId) return;
-          if (otherWs.readyState !== 1) return;
-          otherWs.send(JSON.stringify({ type: "dm_typing", dmKey: dmKey, userId: userId, typing: !!msg.typing }));
-        });
-      });
-      return;
-    }
-
-    if (msg.type === "dm_send") {
-      if (!msg.dmKey || !msg.text) return;
-      var parts = msg.dmKey.split(":");
-
-      // Handle mate DM: dmKey is "mate:mate_xxx"
-      var mateCtx3 = mates.buildMateCtx(userId);
-      if (parts[0] === "mate" && mates.isMate(mateCtx3, parts[1])) {
-        var mate = mates.getMate(mateCtx3, parts[1]);
-        if (!mate) return;
-        // Verify sender is the mate's creator
-        if (mate.createdBy !== userId) return;
-        var message = dm.sendMessage(msg.dmKey, userId, msg.text);
-        ws.send(JSON.stringify({ type: "dm_message", dmKey: msg.dmKey, message: message }));
-        return;
-      }
-
-      // Regular DM: verify sender is a participant
-      if (parts.indexOf(userId) === -1) return;
-      var message = dm.sendMessage(msg.dmKey, userId, msg.text);
-      // Send confirmation to sender
-      ws.send(JSON.stringify({ type: "dm_message", dmKey: msg.dmKey, message: message }));
-      // Broadcast to target user's connections across all projects
-      var targetId = parts[0] === userId ? parts[1] : parts[0];
-      projects.forEach(function (ctx) {
-        ctx.forEachClient(function (otherWs) {
-          if (otherWs === ws) return;
-          if (!otherWs._clayUser || otherWs._clayUser.id !== targetId) return;
-          if (otherWs.readyState !== 1) return;
-          otherWs.send(JSON.stringify({ type: "dm_message", dmKey: msg.dmKey, message: message }));
-        });
-      });
-      // Send push notification to target user
-      if (pushModule && pushModule.sendPushToUser) {
-        var senderName = ws._clayUser ? (ws._clayUser.displayName || ws._clayUser.username || "Someone") : "Someone";
-        var preview = (msg.text || "").substring(0, 140);
-        pushModule.sendPushToUser(targetId, {
-          type: "dm",
-          title: senderName,
-          body: preview,
-          tag: "dm-" + msg.dmKey,
-          dmKey: msg.dmKey,
-        });
-      }
-      return;
-    }
-
-    if (msg.type === "dm_add_favorite") {
-      if (!msg.targetUserId) return;
-      users.removeDmHidden(userId, msg.targetUserId);
-      var updatedFavorites = users.addDmFavorite(userId, msg.targetUserId);
-      var allUsersList = users.getAllUsers().map(function (u) {
-        var p = u.profile || {};
-        return {
-          id: u.id,
-          displayName: p.name || u.displayName || u.username,
-          username: u.username,
-          role: u.role,
-          avatarStyle: p.avatarStyle || "thumbs",
-          avatarSeed: p.avatarSeed || u.username,
-          avatarColor: p.avatarColor || "#7c3aed",
-          avatarCustom: p.avatarCustom || "",
-        };
-      });
-      ws.send(JSON.stringify({
-        type: "dm_favorites_updated",
-        dmFavorites: updatedFavorites,
-        allUsers: allUsersList,
-      }));
-      return;
-    }
-
-    if (msg.type === "dm_remove_favorite") {
-      if (!msg.targetUserId) return;
-      users.addDmHidden(userId, msg.targetUserId);
-      var updatedFavorites = users.removeDmFavorite(userId, msg.targetUserId);
-      ws.send(JSON.stringify({
-        type: "dm_favorites_updated",
-        dmFavorites: updatedFavorites,
-      }));
-      return;
-    }
 
     // --- Mate handlers ---
